+++ R/block.R
@@ -162,11 +162,6 @@ block_exec = function(options) {
   if (keep != 'none' && is.null(options$fig.ext))
     options$fig.ext = dev2ext(options$dev)
 
-  if (!is.null(err.code <- opts_knit$get('stop_on_error'))) {
-    warning('the package option stop_on_error was deprecated;',
-            ' use the chunk option error = ', err.code != 2L, ' instead')
-    options$error = err.code != 2L
-  }
   cache.exists = cache$exists(options$hash, options$cache.lazy)
   evaluate = knit_hooks$get('evaluate')
   # return code with class 'source' if not eval chunks