+++ NEWS.md
@@ -4,6 +4,10 @@
 
 - the minimal version of R required for **knitr** is 3.1.0 now (#1269)
 
+## MINOR CHANGES
+
+- the package option `opts_knit$get('stop_on_error')` has been removed
+
 # CHANGES IN knitr VERSION 1.14
 
 ## NEW FEATURES